+++ packages/query-core/src/queryClient.ts
@@ -3,7 +3,7 @@ import {
   hashCacheKey,
   noop,
   partialMatchKey,
-  resolveStaleTime,
+  resolveQueryValue,
   skipToken,
 } from './utils'
 import { QueryCache } from './queryCache'
@@ -176,7 +176,7 @@ export class QueryClient {
 
     if (
       options.revalidateIfStale &&
-      query.isStaleByTime(resolveStaleTime(defaultedOptions.staleTime, query))
+      query.isStaleByTime(resolveQueryValue(defaultedOptions.staleTime, query))
     ) {
       void this.prefetchQuery(defaultedOptions)
     }
@@ -388,7 +388,7 @@ export class QueryClient {
     const query = this.#queryCache.build(this, defaultedOptions)
 
     const isStale = query.isStaleByTime(
-      resolveStaleTime(defaultedOptions.staleTime, query),
+      resolveQueryValue(defaultedOptions.staleTime, query),
     )
 
     const queryData = isStale
@@ -432,7 +432,7 @@ export class QueryClient {
     const query = this.#queryCache.build(this, defaultedOptions)
 
     return query.isStaleByTime(
-      resolveStaleTime(defaultedOptions.staleTime, query),
+      resolveQueryValue(defaultedOptions.staleTime, query),
     )
       ? query.fetch(defaultedOptions)
       : Promise.resolve(query.state.data as TData)